delete flow/index.js and move content from create/index.js to index.js
